Hune Sprites
Implemented hune sprites. I also think I implemented message dialogs but
I forgot to commit that and also implementing hunes broke the editor I
was working on so whoops I'll have to go clean that up later but I want
to get this committed because i did a lot of work getting this set up
and I'd rather not lose that progress.