Tile Render Hotfix
Okay, I don't know why the shader for flat chunks can't be used if the
other shaders are being used, so I implemented another version of chunks
that uses the shader for walls. And that works, I guess? So I can stop
caring about it for the time being, yay!